% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{aggregate_bootstrap}
\alias{aggregate_bootstrap}
\title{Aggregate bootstrap samples}
\usage{
aggregate_bootstrap(kal, mapping, split_by = "gene_id", column = "tpm",
  aggregate_fun = sum)
}
\arguments{
\item{kal}{a \code{kallisto} object}

\item{mapping}{a data.frame containing the mapping with columns
\code{target_id} and a column specified in 'split_by'}

\item{split_by}{a character string of length one denoting the column in \code{mapping} to split by (such as \code{gene_id})}

\item{aggregate_fun}{a function to aggregate}
}
\value{
a data.frame nrow(mapping) rows that has been aggregated
groupwise using \code{aggregate_fun}
}
\description{
A faster way to aggregate bootstrap samples based off of some mapping.
}

