% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{basic_shrink_fun}
\alias{basic_shrink_fun}
\title{Default Sleuth Variance Shrinkage Estimation}
\usage{
basic_shrink_fun(me_list, ...)
}
\arguments{
\item{me_list}{the list produced by \code{\link{me_model}}.}

\item{...}{advanced options to tweak the shrinkage procedure. See "details"
for more information.}
}
\value{
a modified data.frame with the following additional columns:

\itemize{
  \item \code{'iqr'}: boolean for whether this feature was within the interquartile
  range and used for shrinkage estimation
  \item \code{'failed_ise'}: boolean for whether this feature failed the initial
  round of shrinkage estimation using interpolation from the LOESS curve. This feature's
  shrunken variance was estimated using the 'exact' option for LOESS.
  \item \code{'smooth_sigma_sq'}: the estimated biological variance after the shrinkage
  procedure
}
}
\description{
This function performs the default procedure to shrink variance estimation.
}
\details{
The default procedure does the following steps: it first bins the
features by the mean of each feature's observations. It then takes the
interquartile (25th-75th percentile) of variances within each bin.
Those means and variances are then input into a LOESS to calculate
a curve using the function mean ~ (variance)^1/4. The LOESS is then used
to interpolate the shrunken variance for all features. The larger of the
variance estimated from the bootstraps or the variance from the shrinkage procedure is used.

Advanced options for the shrinkage procedure (these options are passed to
\code{\link{sliding_window_grouping}}):

\itemize{
  \item \code{n_bins}: the number of bins that the data should be split for the sliding window shrinkage
  using the mean-variance curve. The default is 100.
  \item \code{lwr}: the lower range of variances within each bin that should be included for the shrinkage
  procedure. The default is 0.25 (meaning the 25th percentile).
  \item \code{upr}: the upper range of variances within each bin that should be included for the shrinkage
  procedure. The default is 0.75 (meaning the 75th percentile).
}
}
