% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap2mat}
\alias{bootstrap2mat}
\title{Bootstrap to matrix}
\usage{
bootstrap2mat(kal, column = "tpm")
}
\arguments{
\item{kal}{a kallisto object with non-null member \code{bootstrap}}
}
\value{
a matrix with rownames equal to target_id
}
\description{
Takes a \code{kallisto} object and converts the bootstrap results into
a proper \code{matrix}
}

