% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kallisto.R
\name{coef_var}
\alias{coef_var}
\title{kallisto coefficient of variation}
\usage{
coef_var(kal)
}
\arguments{
\item{kal}{a kallisto object}
}
\value{
a data.frame with the coefficient of variation
}
\description{
cv
}

