% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_form_lmm.R
\name{compute_t}
\alias{compute_t}
\title{Compute t-statistics}
\usage{
compute_t(obj, which_var, adjust_sigma = TRUE)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_var}{which variance to use}

\item{adjust_sigma}{if TRUE, adjust sigma}
}
\value{
a data.frame with t-statistics
}
\description{
Compute t-statistics from a "sleuth" object
}

