% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{design_matrix}
\alias{design_matrix}
\title{Extract design matrix}
\usage{
design_matrix(obj, which_model = "full")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_model}{a character string of the model}
}
\value{
the \code{model.matrix} used to fit \code{which_model}
}
\description{
Accessor method for extracting a design matrix from a sleuth object
}

