% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{excluded_ids}
\alias{excluded_ids}
\title{Excluded IDs in Kallisto object}
\usage{
excluded_ids(obj)
}
\arguments{
\item{obj}{a kallisto object}
}
\value{
a character vector containing all of the excluded IDs. This will be
  an empty character vector if the kallisto is the original object.
}
\description{
Returns the excluded IDs if a kallisto object has been subsetted.
Will return an empty character vector if it is the original object.
}
