% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varfit.R
\name{fit_bootstrap_bio}
\alias{fit_bootstrap_bio}
\title{Fit the variance as a function of the mean}
\usage{
fit_bootstrap_bio(sleu)
}
\description{
Fit the variance as a function of the mean, by conditioning on the boostrap
samples.
}

