% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{gene_from_gene}
\alias{gene_from_gene}
\title{Get the gene ID using other gene identifiers}
\usage{
gene_from_gene(obj, gene_colname, gene_name)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{gene_colname}{the name of the column containing 'gene_name'.
This parameter refers to the name of the column that the gene you are searching for appears in.
Check the column names using \code{colnames(obj$target_mapping)}.}

\item{gene_name}{a string containing the name of the gene you are interested in.}
}
\value{
a character vector containing the \code{target_id} of the gene, found under
  \code{obj$gene_column} within \code{obj$target_mapping}.
  If the column name provided is the same as \code{obj$gene_column}, and the
  gene_name used is found, that gene_name will be returned.
}
\description{
Get the \code{target_id} of a gene using other gene identifiers.
The identifiers found under the \code{obj$gene_column} are often
difficult to remember (e.g. ensembl gene ID, ENSG00000111640).
This function allows a user to find that difficult-to-remember
identifier using more-easily-remembered identifiers, such as
gene symbol (e.g. "GAPDH").
}
\examples{
  \dontrun{gene_from_gene(obj, "gene_symbol", "GAPDH")}
}
