% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstrap_summary}
\alias{get_bootstrap_summary}
\title{bootstrap summary}
\usage{
get_bootstrap_summary(obj, target_id, units = "est_counts")
}
\arguments{
\item{obj}{a \code{sleuth} object such that \code{extra_bootstrap_summary = TRUE} inside of \code{\link{sleuth_prep}}.}

\item{target_id}{a character vector of length 1 indicating the target_id (transcript or gene name depending on aggregation mode)}

\item{units}{a character vector of either 'est_counts' or 'tpm' (also requires \code{extra_bootstrap_summary = TRUE} in \code{\link{sleuth_prep}})}
}
\value{
a \code{data.frame} with the summary statistics across all samples for that particular target
}
\description{
Extract the bootstrap summary from a sleuth object that has been initialized in sleuth_prep.
}

