% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstraps}
\alias{get_bootstraps}
\title{Extract bootstrap for a specific transcript}
\usage{
get_bootstraps(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{arguments passed to other functions}
}
\value{
a \code{data.frame} with bootstrap samples
}
\description{
Extract bootstrap for a specific transcript
}

