% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{get_col}
\alias{get_col}
\title{Get a data.frame from all kallisto objects}
\usage{
get_col(obj, ...)
}
\arguments{
\item{obj}{a sleuth object}

\item{col_name}{a column to extract}
}
\value{
a data.frame with columns \code{sample} and column \code{col_name}
}
\description{
Build a data.frame from all kallisto objects given a column name
}

