% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggPlotExpression}
\alias{ggPlotExpression}
\title{Heatmap of expression}
\usage{
ggPlotExpression(exMat, clustRows = TRUE, clustCols = TRUE,
  rowNames = TRUE, colNames = TRUE)
}
\arguments{
\item{exMat}{the expression matrix}

\item{clustRows}{if TRUE, cluster the rows by hierarchical clustering.}

\item{clustCols}{if TRUE, cluster the columns by hierarchical clustering.}

\item{rowNames}{if TRUE, print the row names on the plot}

\item{colNames}{if TRUE, print the column names on the plot}
}
\value{
a ggplot object
}
\description{
Plot all of the points in an expression matrix
}

