% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{is_kallisto_subset}
\alias{is_kallisto_subset}
\title{Is Kallisto Object Subsetted?}
\usage{
is_kallisto_subset(obj)
}
\arguments{
\item{obj}{a sleuth or kallisto object}
}
\value{
a boolean vector with \code{TRUE} if the kallisto object has been subsetted
  (i.e. not all originally analyzed target_ids are represented), or \code{FALSE} if
  the kallisto object contains the full original set of target_ids. For a sleuth object,
  a boolean vector for all of the contained kallisto objects is returned.
}
\description{
This function returns a boolean for whether the kallisto object has been
subsetted or if it contains the original set of target_ids analyzed by kallisto.
If a sleuth object is given, then a vector of booleans for each kallisto object
is returned.
}
