% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{kallisto_table}
\alias{kallisto_table}
\title{Get kallisto abundance table from a sleuth object}
\usage{
kallisto_table(obj, use_filtered = FALSE, normalized = TRUE,
  include_covariates = TRUE)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, return the filtered data}

\item{normalized}{if TRUE, return the normalized data. Otherwise, return
the raw data}

\item{include_covariates}{if TRUE, join the covariates}
}
\value{
a \code{data.frame} with at least columns
}
\description{
Get back the kallisto abundance table from a sleuth object.
}

