% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varfit.R
\name{lf_predict}
\alias{lf_predict}
\title{Fit locfit and predict}
\usage{
lf_predict(x, y, maxk = 1000, ...)
}
\arguments{
\item{x}{predictor variable}

\item{y}{response variable}

\item{maxk}{See \code{\link{locfit::locfit.raw}}}

\item{...}{Addition arguments sent to \code{\link{locfit::locfit.raw}}}
}
\value{
predicted values from fitting y ~ x
}
\description{
Wrapper to fit a locfit model and predict in one call
}

