% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{limma_shrink_fun}
\alias{limma_shrink_fun}
\title{Limma-style variance shrinkage estimation}
\usage{
limma_shrink_fun(me_list, ...)
}
\arguments{
\item{me_list}{the list produced by \code{link{me_model}}}

\item{...}{advanced options that will be passed to the squeezeVar function from limma.
the two advanced options available are:
\itemize{
  \item \code{robust}: boolean for whether the estimation should be made robust to
  outlier variances. The default is \code{FALSE}.
  \item \code{winsor.tail.p}: if robust is \code{TRUE}, this is a numeric vector of
  length 1 or 2 to give the left and right tail proportions of 'x' to Winsorize.
}
please see ?limma::squeezeVar for more information.}
}
\value{
a data.frame taken from me_list$mes_df and modified to have the additional
columns:
  \itemize{
    \item \code{'limma_s2_prior'}: contains the location of the distribution; this
    corresponds to the 's2.prior' output by 'limma::eBayes'.
    \item \code{'limma_df_prior'}: contains the degrees of freedom for the
    prior distribution. This will be the same if robust is \code{FALSE} and
    different for each feature if robust is \code{TRUE}.
    \item \code{'smooth_sigma_sq'}: contains limma's estimated shrunken variances.
    This corresponds to the 's2.post' output by 'limma::eBayes'.
  }
}
\description{
This function uses limma's procedure for variance shrinkage estimation
See ?limma::squeezeVar for more details for how it works and advanced options
available.
}
\seealso{
\code{\link{me_model}} for the measurement error model function
}
