% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_form_lmm.R
\name{lmm}
\alias{lmm}
\title{Simple special case LMM}
\usage{
lmm(X, y, nsamples)
}
\arguments{
\item{X}{the design matrix for the fixed effect}

\item{y}{the response}

\item{nsamples}{the number of biological samples (in LMM terms, the "number
of clusters")}
}
\description{
This is an implementation of the special linear mixed effects model. The
contraints are the design is balanced (within group sizes are equal) and
there is one common random intercept.
}
\details{
Notation is borrowed from "Mixed Models: Theory and Applications with R,
Second Edition" by Eugene Demidenko
}

