% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{log_transform}
\alias{log_transform}
\title{Natural log and offset transformation}
\usage{
log_transform(x, offset = 0.5)
}
\arguments{
\item{x}{numeric that must be >=0. represents an individual observed count (one transcript in one sample).}

\item{offset}{numeric offset to prevent taking the log of 0.}
}
\value{
log(x + offset)
}
\description{
The default transformation function for converting the normalized counts.
}

