% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{me_equal_var}
\alias{me_equal_var}
\title{Measurement error model with equal variances}
\usage{
me_equal_var(obj, pass_filt, xform = function(x) log(x + 0.5))
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{pass_filt}{a data.frame with columns \code{target_id} and
\code{count_filt}. \code{count_filt} is a logical where a target labeled as
FALSE does not pass the filter and thus will not be used in testing}

\item{xform}{a function to transform a matrix}
}
\value{
a \code{data.frame}
}
\description{
Fit the measurement error model with equal variances across samples and
noise
}

