% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{me_model}
\alias{me_model}
\title{Measurement error model}
\usage{
me_model(obj, design, bs_summary, which_var = "obs_counts")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{design}{a design matrix}

\item{bs_summary}{a list from \code{bs_sigma_summary}}

\item{which_var}{"obs_counts" (default) to model estimated counts;
"obs_tpm" to model TPMs}
}
\value{
a list with two items:
   \itemize{
      \item models the list of values returned by lm.fit. At the end of \code{sleuth_fit},
        this object is found here: obj$fits$[[<model_name>]]$models
      \item mes_df this data frame contains all of the important variables that are used for
        variance shrinkage estimation and for testing. The final result of this table
        is found here: obj$fits[[<model_name>]]$summary. This data.frame contains the
        the following columns:
        \itemize{
          \item \code{target_id}: the feature ID (transcript or gene, depending on
          \code{obj$gene_mode})
          \item \code{rss}: the residual sum of squares from the linear model
          \item \code{sigma_sq}: the estimated biological variance from the linear model
          \item \code{mean_obs}: the mean of the transformed observations
          \item \code{var_obs}: the raw variance of the transformed observations
          \item \code{sigma_sq_pmax}: the greater of sigma_sq or 0. This prevents
          negative estimates for the biological variance
        }
   }
}
\description{
Fit the measurement error model across all samples. This is an internal function
called by \code{sleuth_fit}, and is not intended to be used directly by users.
}
