% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{me_model_by_row}
\alias{me_model_by_row}
\title{Measurement error model}
\usage{
me_model_by_row(obj, design, bs_summary)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{design}{a design matrix}

\item{bs_summary}{a list from \code{bs_sigma_summary}}
}
\value{
a list with a bunch of objects that are useful for shrinking
}
\description{
Fit the measurement error model across all samples
}

