% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{melt_bootstrap}
\alias{melt_bootstrap}
\title{Convert kallisto bootstraps into a molten data.frame}
\usage{
melt_bootstrap(kal, column = "tpm", transform = identity)
}
\arguments{
\item{kal}{a kallisto object}

\item{column}{the column to pull out of the kallisto results (default = "tpm")}
}
\value{
a molten data.frame with columns "target_id", "sample" and the selected variable
}
\description{
Melt it!
}

