% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_benchmark.R
\name{merge_results}
\alias{merge_results}
\title{Merge results from expression estimation tools}
\usage{
merge_results(exp_list, exp_labels, oracle)
}
\arguments{
\item{exp_list}{a list of expression results, each which contain columns:
target_id, tpm, est_counts}

\item{exp_labels}{a character vector of the same length as exp_list with
labels for each method}

\item{oracle}{a data.frame with columns: target_id, counts, tpm}
}
\description{
Merge the results from expression estimation tools into one data.table that
is easy to analyze and plot
}

