% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_form_lmm.R
\name{naive_shrink_lmm}
\alias{naive_shrink_lmm}
\title{Naive shrinkage using locfit}
\usage{
naive_shrink_lmm(obj)
}
\arguments{
\item{obj}{a \code{sleuth} object}
}
\value{
a \code{sleuth} object with an extra column in the slot \code{sigma}
}
\description{
The most naive shrinkage you could perform using locfit and the raw
variances from the LMM
}

