% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{new_de_benchmark}
\alias{new_de_benchmark}
\title{Merge results from differential expression tools}
\usage{
new_de_benchmark(de_list, de_labels, oracle)
}
\arguments{
\item{de_list}{a list of de results}

\item{oracle}{a data.frame with the columns target_id, is_de and other optional columns (TBD)}

\item{de_label}{a character vector of the same length as de_list with labels for each method}
}
\description{
Merge results from differential expression tools to benchmark calls. Plots and stuff can be made from this object as well
}

