% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{new_sleuth}
\alias{new_sleuth}
\title{Constructor for a 'sleuth' object}
\usage{
new_sleuth(kal_list, sample_names, condition_names, norm_bootstraps = TRUE)
}
\arguments{
\item{kal_list}{a list of \code{kallisto} objects}

\item{sample_names}{a character vector of \code{length(kal_list)} with
identifiers for the sample.}

\item{condition_names}{a character vector of \code{length(kal_list)} with
identifiers for the condition. These typically refer to the cell type (e.g.
tumor vs not) and should be the same for samples that are part of the same
experimental condition}

\item{norm_boostrap}{if \code{TRUE} use the size factors calculated from the
raw observations to normalize bootstrap values}
}
\value{
a \code{sleuth} object
}
\description{
Conceptually, a sleuth is a pack of kallistos. A 'sleuth' object stores
a pack of kallisto results, and can then intelligently operate them
depending on conditions and sequencing depth.
}

