% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{norm_factors}
\alias{norm_factors}
\title{Normalization factors}
\usage{
norm_factors(mat)
}
\arguments{
\item{mat}{a matrix of estimated counts}
}
\value{
a vector of length \code{ncol(mat)} with a sequencing depth factor
for each sample
}
\description{
More or less the DESeq size factors
}

