% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{normalize_bootstrap}
\alias{normalize_bootstrap}
\title{Normalize bootstrap samples}
\usage{
normalize_bootstrap(kal, tpm_size_factor, est_counts_size_factor)
}
\arguments{
\item{kal}{a kallisto object}

\item{tpm_size_factor}{the size factor (numeric length 1)}

\item{est_counts_size_factor}{the size factor (numeric length 1)}
}
\description{
Normalize by dividing by the "size factor" for each TPM and estimated counts
}

