% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{obs_to_matrix}
\alias{obs_to_matrix}
\title{observations to a matrix}
\usage{
obs_to_matrix(obj, value_name)
}
\arguments{
\item{obj}{is a sleuth object}

\item{value_name}{either "est_counts" or "tpm"}
}
\value{
a matrix with the appropriate names
}
\description{
observations to a matrix
}

