% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{orderByDendrogram}
\alias{orderByDendrogram}
\title{Order by dendrogram}
\usage{
orderByDendrogram(mat)
}
\arguments{
\item{mat}{a matrix where the rows are observations and the columns are different dimensions on the matrix}
}
\value{
a vector of label orderings
}
\description{
Order by dendrogram
}

