% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_benchmark.R
\name{pairwise_cor}
\alias{pairwise_cor}
\title{Compute all pairwise correlations of a unit}
\usage{
pairwise_cor(mres, unit)
}
\arguments{
\item{mres}{a \code{merged_res} object from \code{merged_result}}

\item{a}{character vector of length 1 that is either "tpm" or "est_counts"}
}
\description{
Compute all pairwise correlations of a unit
}

