% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot bootstrap summary}
\usage{
plot_bootstrap(bs_df, ...)
}
\arguments{
\item{bs_df}{a bootstrap summary data.frame from \code{summarize_bootstrap}}

\item{...}{additional arguments to `geom_point`}
}
\value{
a ggplot2 object
}
\description{
Get a d
}

