% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_hmap}
\alias{plot_cluster_hmap}
\title{Plot clustered heatmap}
\usage{
plot_cluster_hmap(transcripts, obj, units = "tpm", trans = "log")
}
\arguments{
\item{transcripts}{a vector of strings containing a list of transcripts to be plotted in a heatmap}

\item{obj}{a \code{sleuth} object}

\item{units}{a string specifying which units to use, either tpm or est_counts}

\item{trans}{a string specifying a function to transform the data by}
}
\value{
a \code{ggplot} object
}
\description{
Plot a clustered heatmap. The clustering is done by the hclust function.
}

