% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fld.kallisto}
\alias{plot_fld.kallisto}
\title{Plot fragment length distribution}
\usage{
\method{plot_fld}{kallisto}(obj)
}
\arguments{
\item{obj}{a kallisto object}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the fragment link the distribution of a specific kallisto run.
}

