% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fld.sleuth}
\alias{plot_fld.sleuth}
\title{Plot fragment length distribution}
\usage{
\method{plot_fld}{sleuth}(obj, sample)
}
\arguments{
\item{obj}{a sleuth object}

\item{sample}{either the sample index (an integer or numeric), or the sample id (a character of length 1)}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the fragment link the distribution of a specific kallisto run.
}

