% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_group_density}
\alias{plot_group_density}
\title{Plot density}
\usage{
plot_group_density(obj, use_filtered = TRUE, units = "est_counts",
  trans = "log", grouping = setdiff(colnames(obj$sample_to_covariates),
  "sample"), offset = 1)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, use filtered data. otherwise use all data}

\item{units}{either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{trans}{a string pointing to a function to use for the transformation.}

\item{grouping}{a string from the columns of \code{sample_to_covariates} in
the sleuth object for which to group and color by}

\item{offset}{the offset so that transformations such as log don't compute
-Inf. If NULL, then will not add an offset}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the density of a some grouping
}

