% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot Loadings and Interpretations}
\usage{
plot_loadings(obj, use_filtered = TRUE, sample = NULL, pc_input = NULL,
  units = "est_counts", pc_count = NULL, scale = FALSE,
  pca_loading_abs = TRUE, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use all data}

\item{sample}{a character string representing the sample to find the loadings of.
If NULL, then loadings will be computed across all samples (and `pc_input` must be specified).}

\item{pc_input}{the principal components to compute loadings for.}

\item{units}{either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{pc_count}{# of PC's}

\item{scale}{scale or not}

\item{pca_loading_abs}{default true, to see all PC's magnitude (recommended)}
}
\value{
a ggplot object
}
\description{
give a principal component, tells you which contribute the most or give a sample, tells you which PC's it contributes to the most
}
