% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Mean variance plot}
\usage{
plot_mean_var(obj)
}
\arguments{
\item{obj}{a "sleuth" object}
}
\value{
a gpplot object with a layer containing points
}
\description{
Basic mean-variance plot
}

