% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pc_variance}
\alias{plot_pc_variance}
\title{Plot PC Variance}
\usage{
plot_pc_variance(obj, use_filtered = TRUE, units = "est_counts",
  pca_number = NULL, scale = FALSE, PC_relative = NULL, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use all data}

\item{units}{either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{pca_number}{user input on how many PC to display, otherwise default is 5}

\item{scale}{determines scaling}

\item{PC_relative}{gives the option to compare subsequent principal components and their contributions}
}
\value{
a ggplot object
}
\description{
Plot PC variances retained by percentage with option to compare specified PC
}

