% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(obj, pc_x = 1L, pc_y = 2L, text_labels = FALSE,
  color_by = NULL, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{pc_x}{integer denoting which principle component to take for the x-axis}

\item{pc_y}{integer denoting which principle component to take for the y-axis}

\item{text_labels}{if TRUE, use text labels instead of points}

\item{color_by}{a variable to color by. if NA, then will leave all as 'black'}

\item{center}{center the samples before running PCA}

\item{scale}{scale the samples before running PCA}

\item{...}{additional arguments passed to \code{\link{geom_point}} or
\code{\link{geom_text}}}
}
\value{
a gpplot object
}
\description{
Plot the principal component analysis (PCA) projections for a set of RNA-Seq experiments.
}

