% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qqnorm}
\alias{plot_qqnorm}
\title{QQ norm plot}
\usage{
plot_qqnorm(obj, which_beta, which_model = "full", sig_level = 0.1,
  point_alpha = 0.2, sig_color = "red", highlight = NULL,
  highlight_color = "green", line_color = "blue")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_beta}{a character string denoting which beta to use for
highlighting the transcript}

\item{which_model}{a character string denoting which model to use for the
test}

\item{sig_level}{the significance level for Fdr}

\item{point_alpha}{the alpha for the points}

\item{sig_color}{what color to make the 'significant' transcripts}

\item{highlight}{a \code{data.frame} with one column, \code{target_id}.
These points will be highlighted in the plot. if \code{NULL}, no points will
be highlighted.}

\item{highlight_color}{the color to highlight points.}

\item{line_color}{what color to make the QQ line}
}
\value{
a \code{ggplot2} object
}
\description{
Create a Q-Q norm plot of the Wald statistics. The x-axis has the
theoretical quantile you would expect from a standard normal distribution.
The y-axis has the observed quantiles. It is a \code{ggplot2} version of
what you would get from \code{\link{qqnorm}} and \code{\link{qqline}}.
}

