% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_density}
\alias{plot_sample_density}
\title{Plot sample density}
\usage{
plot_sample_density(obj, which_sample = obj$sample_to_covariates$sample[1],
  use_filtered = TRUE, units = "est_counts", trans = "log", offset = 1)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_sample}{a character string matching a sample in
\code{obj$sample_to_covariates}}

\item{use_filtered}{if TRUE, use filtered data. Otherwise use all data}

\item{units}{either \code{'est_counts'} (\code{'scaled_reads_per_base'} for gene_mode) or \code{'tpm'}}

\item{trans}{a string pointing to a function to use for the transformation.}

\item{offset}{the offset so that transformations such as log don't compute
-Inf. If NULL, then will not add an offset}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the density of one particular sample
}

