% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Plot sample heatmap}
\usage{
plot_sample_heatmap(obj, use_filtered = TRUE, color_high = "white",
  color_low = "dodgerblue", x_axis_angle = 50,
  annotation_cols = setdiff(colnames(obj$sample_to_covariates),
  "sample"), cluster_bool = TRUE, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use everything}

\item{color_high}{the 'high' color (to label samples that are close)}

\item{color_low}{the 'low' color (to label samples that are far apart)}

\item{x_axis_angle}{the angle at which to put the x-axis labels}

\item{annotation_cols}{a character vector of covariates from
\code{sample_to_covariates} that should be annotated on the heatmap}

\item{cluster_bool}{whether the rows and columns should be hierarchically
clustered. default is \code{TRUE}}

\item{...}{additional arguments to customize the heatmap. passed to
\code{pheatmap}. See ?pheatmap for documentation on additional options.}
}
\value{
a \code{ggplot2} object
}
\description{
Plot sample heatmap using the Jensen-Shannon divergence
}
