% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot}
\usage{
plot_scatter(obj, sample_x = obj$sample_to_covariates$sample[1],
  sample_y = obj$sample_to_covariates$sample[2], offset = 1,
  point_alpha = 0.2, xy_line = TRUE, xy_line_color = "red",
  xtrans = log, ytrans = log, xlim = NULL, ylim = NULL)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{sample_x}{the string corresponding to the sample name in \code{obj$sample_to_covariates}}

\item{sample_y}{same as \code{sample_x} but for the y-axis}

\item{offset}{a linear offset to help deal with zeroes if transforming}

\item{point_alpha}{the alpha on the points}

\item{xy_line}{if TRUE, plot the xy_line}

\item{xy_line_color}{a string denoting the color for the xy line}

\item{xtrans}{a \code{function} denoting the transformation to perform on the x axis}

\item{ytrans}{same as xtrans for the y-axis}

\item{xlim}{a numeric vector of length two denoting the x limits}

\item{ylim}{same as xlim but for the y-axis}
}
\value{
a ggplot object for the scatterplot
}
\description{
Somewhat flexible scatter plot on a sleuth object
}

