% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_vars}
\alias{plot_vars}
\title{Plot technical variance versus observed variance}
\usage{
plot_vars(obj, test = NULL, test_type = "wt", which_model = "full",
  sig_level = 0.1, point_alpha = 0.2, sig_color = "red", xy_line = TRUE,
  xy_line_color = "red", highlight = NULL, highlight_color = "green")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{test}{the name of the test to highlight significant transcripts for}

\item{test_type}{either 'wt' for wald test or 'lrt' for likelihood ratio test}

\item{which_model}{a character string denoting which model to use for the
test}

\item{point_alpha}{the alpha for the points}

\item{xy_line}{if TRUE, plot the xy_line}

\item{xy_line_color}{a string denoting the color for the xy line}

\item{highlight}{a \code{data.frame} with one column, \code{target_id}.
These points will be highlighted in the plot. if \code{NULL}, no points will be highlighted.}

\item{highlight_color}{the color to highlight points.}
}
\value{
a \code{ggplot2} object
}
\description{
Plot technical variance versus observed variance
}

