% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano plot}
\usage{
plot_volcano(obj, test, test_type = "wt", which_model = "full",
  sig_level = 0.1, point_alpha = 0.2, sig_color = "red",
  highlight = NULL)
}
\arguments{
\item{obj}{a  \code{sleuth} object}

\item{test}{a character string denoting which beta to use for
highlighting the transcript}

\item{test_type}{either 'wt' for wald test or 'lrt' for likelihood ratio test. NB: Currently only the wald test is supported.}

\item{which_model}{a character string denoting which model to use for the
test}

\item{sig_level}{the significance level for Fdr}

\item{point_alpha}{the alpha for the points}

\item{sig_color}{what color to make the 'significant' transcripts}

\item{highlight}{a \code{data.frame} with one column, \code{target_id}.
These points will be displayed below in a table.}
}
\value{
a \code{ggplot} object
}
\description{
Plot a volcano plot. A volcano plot is a plot of beta value (regression coefficient)
vs. log(significance). Ideally, it looks like a volcano; more significance typically
results in higher beta
}

