% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.sleuth_model}
\alias{print.sleuth_model}
\title{Print sleuth model}
\usage{
\method{print}{sleuth_model}(obj)
}
\arguments{
\item{obj}{a \code{sleuth_model} object}
}
\value{
obj (invisible)
}
\description{
Print a model that has been fit by sleuth
}

