% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_benchmark.R
\name{read_kallisto_rename}
\alias{read_kallisto_rename}
\title{Read kallisto output for merge_results}
\usage{
read_kallisto_rename(fname)
}
\arguments{
\item{fname}{the path to "expression.txt" from kallisto}
}
\value{
a data.table
}
\description{
Read a kallisto data into a format that place nicely with
\code{merge_results}. Basically equivalent to
\code{read_kallisto("dir")$abundance}, but doesn't do extra stuff with
metadata that \code{read_kallisto_does}
}

