% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_kallisto_tsv}
\alias{read_kallisto_tsv}
\title{Read kallisto plaintext output}
\usage{
read_kallisto_tsv(fname)
}
\arguments{
\item{fname}{the filename for the tsv file}
}
\value{
a \code{kallisto} object (currently missing attributes and
bootstraps)
}
\description{
This function reads kallisto plaintext output. Note, it cannot be used with
sleuth. It also does not read bootstraps since reading plaintext bootstraps
is quite slow.
}

