% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Sample bootstraps}
\usage{
sample_bootstrap(obj, n_samples = 100L)
}
\arguments{
\item{obj}{a \code{kallisto} object}

\item{n_samples}{the number of samples to genenerate}
}
\description{
From a sleuth object, create experiments by randomly sampling bootstraps from each kallisto object
}

