% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{sleuth}
\alias{sleuth}
\title{A sleuth wrapper for fitting the model}
\usage{
sleuth(obj, formula = NULL, fit_name = NULL)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{formula}{currently ignored. a formula specifying the design to fit}

\item{fit_name}{currently ignored. the name to store the fit in the sleuth object}
}
\value{
a sleuth object with updated attributes
}
\description{
TODO: describe
}

