% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sleuth_live_settings}
\alias{sleuth_live_settings}
\title{settings for sleuth_live}
\usage{
sleuth_live_settings(test_type = "wt")
}
\arguments{
\item{test_type}{either 'wt' for the wald test or 'lrt' for the likelihood ratio test}
}
\value{
a named to list with the options and settings
}
\description{
This is a helper function for setting preferences in sleuth live.
Currently, this is somewhat limited, but it will be expanded in the future.
}

