% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{sleuth_load}
\alias{sleuth_load}
\title{load a sleuth object}
\usage{
sleuth_load(file)
}
\arguments{
\item{file}{the file to load}
}
\value{
a \code{sleuth} object
}
\description{
load a sleuth object previously saved with \code{sleuth_save}
}
\seealso{
\code{\link{sleuth_save}}, \code{\link{sleuth_deploy}}
}

