% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihood.R
\name{sleuth_lrt}
\alias{sleuth_lrt}
\title{sleuth likelihood ratio test}
\usage{
sleuth_lrt(obj, null_model, alt_model)
}
\arguments{
\item{obj}{a sleuth object}

\item{null_model}{the null (or "reduced") model}

\item{alt_model}{the alternate (or "full") model}
}
\value{
a data frame with the test statistic, p-value, and target_id
}
\description{
compute the likelihood ratio test for 2 models. this requires that the null
model be nested in the alternate model
}

