% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_prep}
\alias{sleuth_prep}
\title{Constructor for a 'sleuth' object}
\usage{
sleuth_prep(kal_dirs, sample_names, sample_to_covariates, full_model,
  normalize = TRUE, trans = "log", filter_fun = basic_filter,
  annotation = NULL, verbose = TRUE, ...)
}
\arguments{
\item{kal_dirs}{a character vector of length greater than one where each
string points to a kallisto output directory}

\item{sample_to_covariates}{is a \code{data.frame} which contains a mapping
from \code{sample} (a column) to some set of experimental conditions or covariates}

\item{full_model}{is a \code{formula} which explains the full model (design)
of the experiment. It must be consistent with the data.frame supplied in
\code{sample_to_covariates}}

\item{trans}{is a character string of either 'log', or 'cpm'. Currently
ignored.}

\item{filter_fun}{the function to use when filtering.}

\item{...}{additional arguments passed to other functions}
}
\value{
a \code{sleuth} object containing all kallisto samples, metadata,
and summary statistics
}
\description{
Conceptually, a sleuth is a pack of kallistos. A 'sleuth' object stores
a pack of kallisto results, and can then intelligently operate them
depending on conditions and sequencing depth.
}

