% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{sleuth_save}
\alias{sleuth_save}
\title{save a sleuth object}
\usage{
sleuth_save(obj, file)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{the}{location to save the object to}
}
\description{
save a sleuth object
}
\seealso{
\code{\link{sleuth_load}}, \code{\link{sleuth_deploy}}
}

