% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_summarize_bootstrap}
\alias{sleuth_summarize_bootstrap}
\title{Summarize many bootstrap objects}
\usage{
sleuth_summarize_bootstrap(obj, force = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{force}{if \code{FALSE}, then will only compute the summary if it has
not yet been set.}

\item{verbose}{if \code{TRUE}, print verbosely}
}
\value{
a \code{kallisto} object with member \code{bootstrap_summary}
updated and containing a data frame.
}
\description{
Summarize all the bootstrap samples from a kallisto run. The summarized
values are then all put into a data.frame and stored in the \code{sleuth}
object.
}

