% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{sleuth_to_matrix}
\alias{sleuth_to_matrix}
\title{Convert a sleuth object to matrix}
\usage{
sleuth_to_matrix(obj, which_df, which_units)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_df}{character vector of length one. Which type of data to use
("obs_norm" or "obs_raw")}

\item{which_units}{character vector of length one. Which units to use ("tpm"
or "est_counts")}
}
\description{
Convert a sleuth object to a matrix with the condition names.
}

