% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{sleuth_wt}
\alias{sleuth_wt}
\title{Wald test for a sleuth model}
\usage{
sleuth_wt(obj, which_beta, which_model = "full")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_beta}{a character string of length one denoting which beta to
test}

\item{which_model}{a character string of length one denoting which model to
use}
}
\value{
an updated sleuth object
}
\description{
This function computes the Wald test on one specific 'beta' coefficient on
every transcript.
}
\seealso{
\code{\link{models}} to view which models have been fit and which
coefficients can be tested, \code{\link{sleuth_results}} to get back
a data.frame of the results
}

