% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{spread_abundance_by}
\alias{spread_abundance_by}
\title{Spread abundance by a column}
\usage{
spread_abundance_by(abund, var)
}
\arguments{
\item{abund}{the abundance \code{data.frame} from a \code{sleuth} object}

\item{var}{a character array of length one. The variable for which to get "spread" on (e.g. "est_counts").}
}
\description{
Take a data.frame from a sleuth object (e.g. \code{obs_raw}) and cast it
into a matrix where the rows are the target_ids and the columns are the
sample ids. The values are the variable you are "spreading" on.
}

