% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{subset_kallisto}
\alias{subset_kallisto}
\title{Subset a kallisto object}
\usage{
subset_kallisto(obj, ...)
}
\arguments{
\item{obj}{either a sleuth or kallisto object}
}
\value{
a new object with only the subset of target_ids included
}
\description{
Use a vector of target_ids to subset a single kallisto
object or all of kallisto objects in a sleuth object.
}
