% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{transform_fun<-}
\alias{transform_fun<-}
\title{Change sleuth transform counts function}
\usage{
transform_fun(obj, fxn) <- value
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{gene_colname}{the name of the column in which the desired gene apperas gene appears. Once genes have been added to a sleuth
object, you can inspect the genes names present in your sleuth object via \code{obj$target_mapping}, assuming 'obj' is the name of your sleuth object.
This parameter refers to the name of the column that the gene you are searching for appears in. Checkout the column names using \code{names(obj$target_mapping)}}

\item{gene_name}{a string containing the name of the gene you are interested in}
}
\value{
a character vector containing the ID of the gene mapping to 'gene_name'
}
\description{
Replace the transformation function of a sleuth object for estimated counts
}
