% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{transform_fun_counts<-}
\alias{transform_fun_counts<-}
\title{Change sleuth transform counts function}
\usage{
transform_fun_counts(obj, fxn) <- value
}
\description{
Replace the transformation function of a sleuth object for estimated counts
NOTE: if you change the transformation function after having done a fit,
the fit(s) will have to be redone using the new transformation.
}
\examples{
transform_fun_counts(x) <- function(x) identity(x)
}
