% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{transform_status}
\alias{transform_status}
\title{Check Transform Sync Status of Sleuth Fits}
\usage{
transform_status(obj)
}
\arguments{
\item{obj}{a \code{sleuth} object.}
}
\value{
a print out of each fit with the transform sync status.
}
\description{
This method prints out the sync status for all fits of \code{sleuth} object
If the sleuth object's transform function was changed after sleuth_fit was used,
the user will need to redo sleuth_fit for any fits already done.
}

