% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varfit.R
\name{varfit_bootstrap_est}
\alias{varfit_bootstrap_est}
\title{Fit estimated variance from a bunch of bootstraps}
\usage{
varfit_bootstrap_est(obj, pool = TRUE)
}
\arguments{
\item{obj}{a \code{sleuth} object}
}
\value{
a data frame with a summary statistics
}
\description{
Take a bunch of bootstrap samples and compute the Monte Carlo estimate of
the variance using bootstrap samples.
}

