% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_kallisto_hdf5}
\alias{write_kallisto_hdf5}
\title{Write a kallisto object to HDF5}
\usage{
write_kallisto_hdf5(kal, fname, overwrite = TRUE, write_bootstrap = TRUE,
  compression = 6L)
}
\arguments{
\item{kal}{the kallisto object to write out}

\item{fname}{the file name to write out to}

\item{overwrite}{whether the file should be overwritten if it exists}

\item{write_bootstrap}{whether the bootstraps should be written to file}

\item{compression}{an integer between 0 and 7 that indicates the level of compression
to use, with 0 being no compression and 7 being the highest supported by this method.
The default of 6 is a good choice for most applications.}
}
\value{
the kallisto object \code{kal} invisibly.
}
\description{
Write a kallisto object to HDF5.
}
