% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{$<-.sleuth}
\alias{$<-.sleuth}
\title{Extend internal '$<-' for sleuth object}
\usage{
\method{$}{sleuth}(obj, name) <- value
}
\description{
This extension is mainly to address case where
transform_fun is changed by user.
This function informs user that the fits need to be redone
and updates those fits.
Otherwise it acts normally.
}
\examples{
obj$transform_fun <- function(x) log2(x+0.5)
}

