% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstraps}
\alias{get_bootstraps}
\title{Extract bootstrap for a specific transcript}
\usage{
get_bootstraps(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{arguments passed to other functions}
}
\value{
a \code{data.frame} with bootstrap samples
}
\description{
Extract bootstrap for a specific transcript.
CURRENTLY DEPRECATED: Will probably be re implemented in the next version.
Currently not working because of a complete rewrite of the bootstrap code.
If you are interested in getting the bootstraps, you can manually write some code
using `read_kallisto()`. Please make sure to comment in the user group if you are using this function.
}

