% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{new_sleuth}
\alias{new_sleuth}
\title{Constructor for a 'sleuth' object}
\usage{
new_sleuth(kal_list, sample_to_condition, design, norm_bootstraps = TRUE,
  verbose = TRUE)
}
\arguments{
\item{kal_list}{a list of \code{kallisto} objects}

\item{sample_to_condition}{is a \code{data.frame} which contains a mapping from \code{sample} (a column) to some set of experimental conditions}

\item{design}{is a \code{formula} which explains the full model (design) of the experiment}

\item{norm_boostrap}{if \code{TRUE} use the size factors calculated from the
raw observations to normalize bootstrap values}
}
\value{
a \code{sleuth} object
}
\description{
Conceptually, a sleuth is a pack of kallistos. A 'sleuth' object stores
a pack of kallisto results, and can then intelligently operate them
depending on conditions and sequencing depth.
}

