% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot bootstrap summary}
\usage{
plot_bootstrap(obj, transcript, units = "est_counts",
  color_by = setdiff(colnames(obj$sample_to_covariates), "sample"),
  x_axis_angle = 50)
}
\arguments{
\item{bs_df}{a bootstrap summary data.frame from \code{summarize_bootstrap}}

\item{...}{additional arguments to `geom_point`}
}
\value{
a ggplot2 object
}
\description{
Get a d
}

