% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot bootstrap summary}
\usage{
plot_bootstrap(obj, target_id, units = "est_counts",
  color_by = setdiff(colnames(obj$sample_to_covariates), "sample"),
  x_axis_angle = 50, divide_groups = TRUE)
}
\arguments{
\item{obj}{a sleuth object that contains a bootstrap summary (see \code{\link{get_bootstrap_summary}})}

\item{target_id}{a character vector of length 1 indicating the target_id (transcript or gene name depending on aggregation mode)}

\item{units}{a character vector of either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{color_by}{a column in the sample to covariates to color by}

\item{x_axis_angle}{the angle of the x-axis labels}
}
\value{
a ggplot2 object
}
\description{
Plot the normalized bootstraps across all samples.
}

