% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Mean-variance relationship}
\usage{
plot_mean_var(obj, which_model = "full", point_alpha = 0.4,
  point_size = 2, point_colors = c("black", "dodgerblue"),
  smooth_alpha = 1, smooth_size = 0.75, smooth_color = "red")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_model}{which fit to use}

\item{point_alpha}{the alpha of the points (0, 1)}
}
\value{
a \code{ggvis} object
}
\description{
Plot the mean-variance relationship using ggvis
}

