% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Mean-variance relationship}
\usage{
plot_mean_var(obj, which_model = "full", point_alpha = 0.4,
  point_size = 2, point_colors = c("black", "dodgerblue"),
  smooth_alpha = 1, smooth_size = 0.75, smooth_color = "red")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_model}{which fit to use}

\item{point_alpha}{the alpha (opacity) of the points (0, 1)}

\item{point_size}{the size of the points}

\item{smooth_alpha}{the alpha (opacity) of the line}

\item{smooth_size}{the size of the line}

\item{smooth_color}{the color of the smooth line}
}
\value{
a \code{ggplot} object
}
\description{
Plot the mean-variance relationship that for a set of samples. Each
dot in the plot corresponds to one transcript. The blue dots represent the transcripts that
are used in the shrinkage estimation. The curve is the smoothed fit.
}
\details{
The x-axis values are the mean abundance estimates pooled across all samples. The
y-axis values are the 'biological' variance, estimated after the technical variance has
been removed.
}

