% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean_var}
\alias{plot_mean_var}
\title{Mean-variance relationship}
\usage{
plot_mean_var(obj, which_model = "full", point_alpha = 0.4,
  point_size = 2, point_colors = c("black", "dodgerblue"),
  smooth_alpha = 1, smooth_size = 0.75, smooth_color = "red")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_model}{which fit to use}

\item{point_alpha}{the alpha (opacity) of the points (0, 1)}

\item{point_size}{the size of the points}

\item{smooth_alpha}{the alpha (opacity) of the line}

\item{smooth_size}{the size of the line}

\item{smooth_color}{the color of the smooth line}
}
\value{
a \code{ggplot} object
}
\description{
Plot the mean-variance relationship of transcripts as modeled in \code{sleuth}. Each
dot represents a transcript. The blue dots represent the transcripts that
are used in the shrinkage estimation. The fitted curve represents the smooth fit.
}
\details{
The x-axis represents the mean expression of transcripts pooled across all samples. The
y-axis represents the 'biological' variance after the technical variance has
been removed.
}

