% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(obj, pc_x = 1L, pc_y = 2L, use_filtered = TRUE,
  units = "est_counts", text_labels = FALSE, color_by = NULL,
  point_size = 3, point_alpha = 0.8, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{pc_x}{integer denoting the principle component to use for the x-axis}

\item{pc_y}{integer denoting the principle component to use for the y-axis}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use all data}

\item{units}{either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{text_labels}{if TRUE, use text labels instead of points}

\item{color_by}{a variable to color by. if NA, then will leave all as 'black'}
}
\value{
a gpplot object
}
\description{
Plot projections of samples onto the principal components for a set of RNA-Seq experiments
}

