% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Plot sample heatmap}
\usage{
plot_sample_heatmap(obj, use_filtered = TRUE, color_high = "white",
  color_low = "dodgerblue", x_axis_angle = 50)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use everything}

\item{color_high}{the 'high' color}

\item{color_low}{the 'low' color}

\item{x_axis_angle}{the angle at which to put the x-axis labels}
}
\value{
a \code{ggplot2} object
}
\description{
Plot sample heatmap using the Jensen-Shannon divergence
}

