% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Sample to sample scatter plot}
\usage{
plot_scatter(obj, sample_x = obj$sample_to_covariates$sample[1],
  sample_y = obj$sample_to_covariates$sample[2], use_filtered = TRUE,
  units = "est_counts", offset = 1, point_alpha = 0.2, xy_line = TRUE,
  xy_line_color = "red", trans = "log", xlim = NULL, ylim = NULL)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{sample_x}{the string corresponding to the sample name in \code{obj$sample_to_covariates}}

\item{sample_y}{same as \code{sample_x} but for the y-axis}

\item{use_filtered}{if TRUE, use filtered data. otherwise, use all data}

\item{units}{either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'}

\item{offset}{a linear offset to help deal with zeroes if transforming the abundances}

\item{point_alpha}{the alpha on the points}

\item{xy_line}{if TRUE, plot the xy_line}

\item{xy_line_color}{a string denoting the color for the xy line}

\item{trans}{a string pointing to a function to use for the transformation.
This function must exist in the global namespace. This means you should be
able to call \code{eval('myfun')} and get a function back.}

\item{xlim}{a numeric vector of length two denoting the x limits}

\item{ylim}{same as xlim but for the y-axis}
}
\value{
a ggplot object for the scatterplot
}
\description{
Make a scatter plot of transcripts from two samples. to assess correlation
}

