% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_transcript_heatmap}
\alias{plot_transcript_heatmap}
\title{Plot clustered heatmap}
\usage{
plot_transcript_heatmap(obj, transcripts, units = "tpm", trans = "log",
  offset = 1)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{transcripts}{a vector of strings containing a list of transcripts to be plotted in a heatmap}

\item{units}{a string specifying which units to use, either tpm or est_counts (scaled_reads_per_base for gene_mode)}

\item{trans}{a string specifying a function to transform the data by}
}
\value{
a \code{ggplot} object
}
\description{
Plot a clustered heatmap. The clustering is done by the hclust function.
}

