% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_kallisto}
\alias{read_kallisto}
\title{Read kallisto output}
\usage{
read_kallisto(path, read_bootstrap = TRUE, max_bootstrap = NULL)
}
\arguments{
\item{read_bootstrap}{if \code{TRUE}, bootstraps will be read (h5 only)}

\item{max_bootstrap}{an integer denoting the number of bootstraps to read.
if \code{NULL} read everything available (h5 only)}

\item{either}{the kallisto directory name or the file name of a h5 or tsv output from kallisto}
}
\value{
a \code{kallisto} object
}
\description{
This is a general driver function to read kallisto output. It can read either
H5 or tsv.
}

