% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{sleuth_fit}
\alias{sleuth_fit}
\title{Fit a measurement error model}
\usage{
sleuth_fit(obj, formula = NULL, fit_name = NULL, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{formula}{a formula specifying the design to fit}

\item{fit_name}{the name to store the fit in the sleuth
object}

\item{...}{additional arguments passed to \code{sliding_window_grouping} and
\code{shrink_df}}
}
\value{
a sleuth object with updated attributes
}
\description{
This function is a wrapper for fitting a measurement error model using
\code{sleuth}.
}
\seealso{
\code{\link{sleuth_prep}} for creating a sleuth object,
\code{\link{sleuth_test}} to test whether a coefficient is zero
}

