% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_gene_table}
\alias{sleuth_gene_table}
\title{Create a gene table from a sleuth object}
\usage{
sleuth_gene_table(obj, test, test_type = "lrt", which_model = "full",
  which_group = "ens_gene")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{test}{a character string denoting which beta to use}

\item{test_type}{either 'wt' for Wald test or 'lrt' for likelihood ratio test}

\item{which_model}{a character string denoting which model to use}

\item{which_group}{a character string denoting which gene group to use}
}
\value{
a \code{data.frame} containing gene names, transcript names, and significance
}
\description{
Take in a sleuth object with added genes, and return a data table in which genes
list the most significant transcript mapping to themselves
}

