% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{sleuth_live}
\alias{sleuth_live}
\title{Interactive sleuth visualization with Shiny}
\usage{
sleuth_live(obj, ...)
}
\arguments{
\item{obj}{a \code{sleuth} object already processed and has run
\code{\link{sleuth_fit}} and \code{\link{sleuth_test}}}

\item{...}{additional parameters sent to plotting functions}
}
\value{
a \code{\link{shinyApp}} result
}
\description{
Interactive sleuth visualization with Shiny. To exit, type \code{ESC} in R.
}
\seealso{
\code{\link{sleuth_fit}}, \code{\link{sleuth_test}}
}

