% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{sleuth_lrt}
\alias{sleuth_lrt}
\title{sleuth likelihood ratio test}
\usage{
sleuth_lrt(obj, null_model, alt_model)
}
\arguments{
\item{obj}{a sleuth object}

\item{null_model}{the null (or "reduced") model}

\item{alt_model}{the alternate (or "full") model}
}
\value{
an updated sleuth object with a likelihood ratio test computation
}
\description{
compute the likelihood ratio test for 2 models. this requires that the null
model be nested in the alternate model
}
\seealso{
\code{\link{models}} to view which models have been fit and which
coefficients can be tested, \code{\link{sleuth_results}} to get back
a data.frame of the results
}

