% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_prep}
\alias{sleuth_prep}
\title{Constructor for a 'sleuth' object}
\usage{
sleuth_prep(kal_dirs, sample_to_covariates, full_model,
  filter_fun = basic_filter, target_mapping = NULL, ...)
}
\arguments{
\item{kal_dirs}{a character vector of length greater than one where each
string points to a kallisto output directory}

\item{sample_to_covariates}{is a \code{data.frame} which contains a mapping
from \code{sample} (a column) to some set of experimental conditions or
covariates. The column \code{sample} should be in the same order as the
corresponding entry in \code{kal_dirs}}

\item{full_model}{is a \code{formula} which explains the full model (design)
of the experiment. It must be consistent with the data.frame supplied in
\code{sample_to_covariates}}

\item{filter_fun}{the function to use when filtering.}

\item{target_mapping}{a \code{data.frame} that has at least one column
'target_id' and others that denote the mapping for each target. if it is not
\code{NULL}, \code{target_mapping} is joined with many outputs where it
might be useful. For example, you might have columns 'target_id',
'ensembl_gene' and 'entrez_gene' to denote different transcript to gene
mappings.}

\item{...}{additional arguments passed to the filter function}
}
\value{
a \code{sleuth} object containing all kallisto samples, metadata,
and summary statistics
}
\description{
A sleuth is a group of kallistos. Borrowing this terminology, a 'sleuth' object stores
a group of kallisto results, and can then operate on them while
accounting for covariates, sequencing depth, technical and biological
variance.
}
\seealso{
\code{\link{sleuth_fit}} to fit a model, \code{\link{sleuth_test}} to
test whether a coeffient is zero
}

