% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_prep}
\alias{sleuth_prep}
\title{Constructor for a 'sleuth' object}
\usage{
sleuth_prep(sample_to_covariates, full_model, filter_fun = basic_filter,
  target_mapping = NULL, max_bootstrap = NULL,
  norm_fun_counts = norm_factors, norm_fun_tpm = norm_factors, ...)
}
\arguments{
\item{sample_to_covariates}{is a \code{data.frame} which contains a mapping
from \code{sample} (a column) to some set of experimental conditions or
covariates. The column \code{path} is also required, which is a character
vector where each element points to the corresponding kallisto output directory. The column
\code{sample} should be in the same order as the corresponding entry in
\code{path}.}

\item{full_model}{is a \code{formula} which explains the full model (design)
of the experiment OR a design matrix. It must be consistent with the data.frame supplied in
\code{sample_to_covariates}.}

\item{filter_fun}{the function to use when filtering.}

\item{target_mapping}{a \code{data.frame} that has at least one column
'target_id' and others that denote the mapping for each target. if it is not
\code{NULL}, \code{target_mapping} is joined with many outputs where it
might be useful. For example, you might have columns 'target_id',
'ensembl_gene' and 'entrez_gene' to denote different transcript to gene
mappings.}

\item{max_bootstrap}{maximum number of bootstrap values to read for each
transcript.}

\item{norm_fun_counts}{a function to perform between sample normalization on the estimated counts.}

\item{norm_fun_tpm}{a function to perform between sample normalization on the TPM}

\item{...}{additional arguments passed to the filter function}
}
\value{
a \code{sleuth} object containing all kallisto samples, metadata,
and summary statistics
}
\description{
A sleuth is a group of kallistos. Borrowing this terminology, a 'sleuth' object stores
a group of kallisto results, and can then operate on them while
accounting for covariates, sequencing depth, technical and biological
variance.
}
\seealso{
\code{\link{sleuth_fit}} to fit a model, \code{\link{sleuth_test}} to
test whether a coeffient in the model is zero
}

