% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sleuth_results}
\alias{sleuth_results}
\title{Extract Wald or Likelihood Ratio test results from a sleuth object}
\usage{
sleuth_results(obj, test, test_type = "wt", which_model = "full",
  rename_cols = TRUE, show_all = TRUE,
  pval_aggregate = obj$pval_aggregate)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{test}{a character string denoting the test to extract. Possible tests can be found by using \code{models(obj)}.}

\item{test_type}{'wt' for Wald test or 'lrt' for Likelihood Ratio test.}

\item{which_model}{a character string denoting the model. If extracting a wald test, use the model name.
Not used if extracting a likelihood ratio test.}

\item{rename_cols}{if \code{TRUE} will rename some columns to be shorter and
consistent with vignette}

\item{show_all}{if \code{TRUE} will show all transcripts (not only the ones
passing filters). The transcripts that do not pass filters will have
\code{NA} values in most columns.}

\item{pval_aggregate}{if \code{TRUE} and both \code{target_mapping} and \code{aggregation_column} were provided,
to \code{sleuth_prep}, use lancaster's method to aggregate p-values by the \code{aggregation_column}.}
}
\value{
a \code{data.frame} with the following columns:

target_id: transcript name, e.g. "ENSXX#####" (dependent on the transcriptome used in kallisto)

...: if there is a target mapping data frame, all of the annotations columns are added from \code{obj$target_mapping} before the other columns.

pval: p-value of the chosen model

qval: false discovery rate adjusted p-value, using Benjamini-Hochberg (see \code{\link{p.adjust}})

test_stat (LRT only): Chi-squared test statistic (likelihood ratio test). Only seen with Likelihood Ratio test results.

rss (LRT only): the residual sum of squares under the "null model". Only seen with Likelihood Ratio test results.

degrees_free (LRT only): the degrees of freedom (equal to difference between the two models). Only seen with Likelihood Ratio test results.

b (Wald only): 'beta' value (effect size). Technically a biased estimator of the fold change. Only seen with Wald test results.

se_b (Wald only): standard error of the beta. Only seen with Wald test results.

mean_obs: mean of natural log counts of observations

var_obs: variance of observation

tech_var: technical variance of observation from the bootstraps (named 'sigma_q_sq' if rename_cols is \code{FALSE})

sigma_sq: raw estimator of the variance once the technical variance has been removed

smooth_sigma_sq: smooth regression fit for the shrinkage estimation

final_simga_sq: max(sigma_sq, smooth_sigma_sq); used for covariance estimation of beta
  (named 'smooth_sigma_sq_pmax' if rename_cols is \code{FALSE})
}
\description{
This function extracts Wald or Likelihood Ratio test results from a sleuth object.
}
\examples{
models(sleuth_obj) # for this example, assume the formula is ~condition,
                     and a coefficient is IP
results_table <- sleuth_results(sleuth_obj, 'conditionIP')
}
\seealso{
\code{\link{sleuth_wt}} and \code{\link{sleuth_lrt}} to compute tests, \code{\link{models}} to
view which models, \code{\link{tests}} to view which tests were performed (and can be extracted)
}
