% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{sleuth_results}
\alias{sleuth_results}
\title{Extract Wald test results from a sleuth object}
\usage{
sleuth_results(obj, which_beta, which_model = "full", rename_cols = TRUE,
  show_all = TRUE)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_beta}{a character string denoting which coefficient test to
extract}

\item{which_model}{a character string denoting which model to extract}

\item{rename_cols}{if \code{TRUE} will rename some columns to be shorter and
consistent with vignette}

\item{show_all}{if \code{TRUE} will show all transcripts (not only the ones
passing filters). The transcripts that do not pass filters will have
\code{NA} values in most columns.}
}
\value{
a \code{data.frame}
}
\description{
This function extracts Wald test results from a sleuth object.
}
\seealso{
\code{\link{wald_test}} to compute tests, \code{\link{models}} to
view which models and betas have been tested
}

