% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{sleuth_results}
\alias{sleuth_results}
\title{Extract Wald test results from a sleuth object}
\usage{
sleuth_results(obj, test, test_type = "wt", which_model = "full",
  rename_cols = TRUE, show_all = TRUE)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{test}{a character denoting the test to extract}

\item{which_model}{a character string denoting the model. If extracting a wald test, use the model name. If extracting a likelihood ratio test, use 'lrt'.}

\item{rename_cols}{if \code{TRUE} will rename some columns to be shorter and
consistent with vignette}

\item{show_all}{if \code{TRUE} will show all transcripts (not only the ones
passing filters). The transcripts that do not pass filters will have
\code{NA} values in most columns.}
}
\value{
a \code{data.frame}
}
\description{
This function extracts Wald test results from a sleuth object.
}
\seealso{
\code{\link{sleuth_test}} to compute tests, \code{\link{models}} to
view which models, \code{\link{tests}} to view which tests were performed (and can be extracted)
}

