% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sleuth_to_matrix}
\alias{sleuth_to_matrix}
\title{Convert a sleuth object to matrix}
\usage{
sleuth_to_matrix(obj, which_df, which_units)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_df}{character vector of length one. Which type of data to use
("obs_norm" or "obs_raw")}

\item{which_units}{character vector of length one. Which units to use ("tpm"
or "est_counts")}
}
\value{
a \code{list} with an attribute 'data', which contains a matrix of target_ids
        and transcript expression in \code{which_units}
}
\description{
Convert a sleuth object to a matrix with the condition names.
}
\examples{
sleuth_matrix <- sleuth_to_matrix(sleuth_obj, 'obs_norm', 'tpm')
head(sleuth_matrix$data) # look at first 5 transcripts, sorted by name
}

