% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{transform_fun<-}
\alias{transform_fun<-}
\title{Change sleuth transform function}
\usage{
transform_fun(obj, fxn) <- value
}
\description{
Replace the transformation function of a sleuth object
}
\details{
NOTE: if you change the transformation function after having done a fit,
the fit(s) will have to be redone using the new transformation.
}
\examples{
transform_fun(x) <- function(x) log2(x+0.5)
}

