% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_kallisto_hdf5}
\alias{write_kallisto_hdf5}
\title{Write a kallisto object to HDF5}
\usage{
write_kallisto_hdf5(kal, fname, overwrite = TRUE)
}
\arguments{
\item{kal}{the kallisto object to write out}

\item{fname}{the file name to write out to}
}
\value{
the kallisto object \code{kal} invisibly.
}
\description{
Write a kallisto object to HDF5.
}

